Add ascii.toString(int) and rename std.fs
This adds the method ascii.toString(int) for converting integers into
strings.  An interesting question is where this method should be
located.  For example, should it be in ascii or in integer?

This also renames std.fs to be std.filesystem.  This makes sense, though
of course the name is rather long.  Providing some kind of ability to do
this:

import std.filesystem as fs

So we could then refer to e.g. fs:File, etc would be helpful.