Add (incomplete) constraints for string type
This adds a set of incomplete constraints for describing utf8 strings.
The missing parts are to do with the length of a string.  For example,
the last byte cannot be the start of a two byte codepoint.

At the moment, there is a problem in that the code seems correct but is
not compiling due to some bug related to property expansion in the JVM
backend.